完善规范文档内容、代码逻辑

+++ integration-tests/source-test.go
@@ -14,6 +14,183 @@ var (
 	ConfigFile = filepath.Join(folderutil.AppConfigDirOrDefault(".uncover-config", "uncover"), "provider-config.yaml")
 )
 
+type qianxunSpiderTestcases struct {
+}
+
+func (q qianxunSpiderTestcases) Execute() error {
+	results, err := testutils.RunUncoverAndGetResults(debug, "-qianxun-spider", "baidu.com")
+	if err != nil {
+		return err
+	}
+	return expectResultsGreaterThanCount(results, 0)
+}
+
+type rapiddnsSpiderTestcases struct {
+}
+
+func (r rapiddnsSpiderTestcases) Execute() error {
+	results, err := testutils.RunUncoverAndGetResults(debug, "-rapiddns-spider", "baidu.com")
+	if err != nil {
+		return err
+	}
+	return expectResultsGreaterThanCount(results, 0)
+}
+
+type baiduSpiderTestcases struct {
+}
+
+func (b baiduSpiderTestcases) Execute() error {
+	results, err := testutils.RunUncoverAndGetResults(debug, "-baidu-spider", "baidu.com")
+	if err != nil {
+		return err
+	}
+	return expectResultsGreaterThanCount(results, 0)
+}
+
+type shodanidbTestcases struct {
+}
+
+func (s shodanidbTestcases) Execute() error {
+	results, err := testutils.RunUncoverAndGetResults(debug, "-shodan-idb", "baidu.com")
+	if err != nil {
+		return err
+	}
+	return expectResultsGreaterThanCount(results, 0)
+}
+
+type yahooSpiderTestcases struct {
+}
+
+func (y yahooSpiderTestcases) Execute() error {
+	results, err := testutils.RunUncoverAndGetResults(debug, "-yahoo-spider", "baidu.com")
+	if err != nil {
+		return err
+	}
+	return expectResultsGreaterThanCount(results, 0)
+}
+
+type sitedossierSpiderTestcases struct {
+}
+
+func (s sitedossierSpiderTestcases) Execute() error {
+	results, err := testutils.RunUncoverAndGetResults(debug, "-sitedossier-spider", "baidu.com")
+	if err != nil {
+		return err
+	}
+	return expectResultsGreaterThanCount(results, 0)
+}
+
+type bingSpiderTestcases struct {
+}
+
+func (b bingSpiderTestcases) Execute() error {
+	results, err := testutils.RunUncoverAndGetResults(debug, "-bing-spider", "baidu.com")
+	if err != nil {
+		return err
+	}
+	return expectResultsGreaterThanCount(results, 0)
+}
+
+type chinazSpiderTestcases struct {
+}
+
+func (c chinazSpiderTestcases) Execute() error {
+	results, err := testutils.RunUncoverAndGetResults(debug, "-chinaz-spider", "baidu.com")
+	if err != nil {
+		return err
+	}
+	return expectResultsGreaterThanCount(results, 0)
+}
+
+type googleSpiderTestcases struct {
+}
+
+func (g googleSpiderTestcases) Execute() error {
+	results, err := testutils.RunUncoverAndGetResults(debug, "-google-spider", "baidu.com")
+	if err != nil {
+		return err
+	}
+	return expectResultsGreaterThanCount(results, 0)
+}
+
+type ip138SpiderTestcases struct {
+}
+
+func (i ip138SpiderTestcases) Execute() error {
+	results, err := testutils.RunUncoverAndGetResults(debug, "-ip138-spider", "baidu.com")
+	if err != nil {
+		return err
+	}
+	return expectResultsGreaterThanCount(results, 0)
+}
+
+type anubisSpiderTestcases struct {
+}
+
+func (a anubisSpiderTestcases) Execute() error {
+	//TODO implement me
+	results, err := testutils.RunUncoverAndGetResults(debug, "-anubis-spider", "baidu.com")
+	if err != nil {
+		return err
+	}
+	return expectResultsGreaterThanCount(results, 0)
+}
+
+type publicwwwTestcases struct {
+}
+
+func (p publicwwwTestcases) Execute() error {
+	token := os.Getenv("PUBLICWWW_API_KEY")
+	if token == "" {
+		return errors.New("missing publicwww api key")
+	}
+	publicwwwToken := fmt.Sprintf(`publicwww: [%s]`, token)
+	_ = os.WriteFile(ConfigFile, []byte(publicwwwToken), os.ModePerm)
+	defer os.RemoveAll(ConfigFile)
+	results, err := testutils.RunUncoverAndGetResults(debug, "-publicwww", "'Grafana'")
+	if err != nil {
+		return err
+	}
+	return expectResultsGreaterThanCount(results, 0)
+}
+
+type binaryTestcases struct {
+}
+
+func (b binaryTestcases) Execute() error {
+	token := os.Getenv("BINARYEDGE_API_KEY")
+	if token == "" {
+		return errors.New("missing binaryedge api key")
+	}
+	binaryToken := fmt.Sprintf(`binaryedge: [%s]`, token)
+	_ = os.WriteFile(ConfigFile, []byte(binaryToken), os.ModePerm)
+	defer os.RemoveAll(ConfigFile)
+	//TODO implement me
+	results, err := testutils.RunUncoverAndGetResults(debug, "-binaryedge", "baidu.com")
+	if err != nil {
+		return err
+	}
+	return expectResultsGreaterThanCount(results, 0)
+}
+
+type githubTestcases struct {
+}
+
+func (g githubTestcases) Execute() error {
+	token := os.Getenv("GITHUB_TOKEN")
+	if token == "" {
+		return errors.New("missing github api key")
+	}
+	githubToken := fmt.Sprintf(`github: [%s]`, token)
+	_ = os.WriteFile(ConfigFile, []byte(githubToken), os.ModePerm)
+	defer os.RemoveAll(ConfigFile)
+	results, err := testutils.RunUncoverAndGetResults(debug, "-github", "baidu.com")
+	if err != nil {
+		return err
+	}
+	return expectResultsGreaterThanCount(results, 0)
+}
+
 type fofaSpiderTestcases struct{}
 
 func (f fofaSpiderTestcases) Execute() error {
@@ -28,12 +205,16 @@ func (f fofaSpiderTestcases) Execute() error {
 type censysTestcases struct{}
 
 func (h censysTestcases) Execute() error {
-	token := os.Getenv("CENSYS_API_KEY")
-	if token == "" {
-		return errors.New("missing censys api key")
+	id := os.Getenv("CENSYS_API_ID")
+	if id == "" {
+		return errors.New("missing censys api id")
 	}
-	censysToken := fmt.Sprintf(`censys: [%s]`, token)
-	_ = os.WriteFile(ConfigFile, []byte(censysToken), 0644)
+	secret := os.Getenv("CENSYS_API_SECRET")
+	if secret == "" {
+		return errors.New("missing censys api secret")
+	}
+	censysToken := fmt.Sprintf(`censys: [%s:%S]`, id, secret)
+	_ = os.WriteFile(ConfigFile, []byte(censysToken), os.ModePerm)
 	defer os.RemoveAll(ConfigFile)
 	results, err := testutils.RunUncoverAndGetResults(debug, "-censys", "'services.software.vendor=Grafana'")
 	if err != nil {
@@ -50,7 +231,7 @@ func (h shodanTestcases) Execute() error {
 		return errors.New("missing shodan api key")
 	}
 	shodanToken := fmt.Sprintf(`shodan: [%s]`, token)
-	_ = os.WriteFile(ConfigFile, []byte(shodanToken), 0644)
+	_ = os.WriteFile(ConfigFile, []byte(shodanToken), os.ModePerm)
 	defer os.RemoveAll(ConfigFile)
 	results, err := testutils.RunUncoverAndGetResults(debug, "-shodan", "'title:\"Grafana\"'")
 	if err != nil {
@@ -75,7 +256,8 @@ func (h zoomeyeTestcases) Execute() error {
 		return errors.New("missing zoomeye api key")
 	}
 	zoomeyeToken := fmt.Sprintf(`zoomeye: [%s]`, token)
-	_ = os.WriteFile(ConfigFile, []byte(zoomeyeToken), 0644)
+
+	_ = os.WriteFile(ConfigFile, []byte(zoomeyeToken), os.ModePerm)
 	defer os.RemoveAll(ConfigFile)
 	results, err := testutils.RunUncoverAndGetResults(debug, "-zoomeye", "'app:\"Atlassian JIRA\"'")
 	if err != nil {
@@ -87,12 +269,16 @@ func (h zoomeyeTestcases) Execute() error {
 type fofaTestcases struct{}
 
 func (h fofaTestcases) Execute() error {
-	token := os.Getenv("FOFA_API_KEY")
+	token := os.Getenv("FOFA_KEY")
 	if token == "" {
 		return errors.New("missing fofa api key")
 	}
-	fofaToken := fmt.Sprintf(`fofa: [%s]`, token)
-	_ = os.WriteFile(ConfigFile, []byte(fofaToken), 0644)
+	email := os.Getenv("FOFA_EMAIL")
+	if email == "" {
+		return errors.New("missing fofa email")
+	}
+	fofaToken := fmt.Sprintf(`fofa: [%s:%s]`, token, email)
+	_ = os.WriteFile(ConfigFile, []byte(fofaToken), os.ModePerm)
 	defer os.RemoveAll(ConfigFile)
 	results, err := testutils.RunUncoverAndGetResults(debug, "-fofa", "'app=Grafana'")
 	if err != nil {
@@ -101,25 +287,25 @@ func (h fofaTestcases) Execute() error {
 	return expectResultsGreaterThanCount(results, 0)
 }
 
-// type hunterTestcases struct{}
+type hunterTestcases struct{}
 
-// func (h hunterTestcases) Execute() error {
-// 	results, err := testutils.RunUncoverAndGetResults(debug, "-hunter", "'Grafana'")
-// 	if err != nil {
-// 		return err
-// 	}
-// 	return expectResultsGreaterOrEqualToCount(results, 0)
-// }
+func (h hunterTestcases) Execute() error {
+	results, err := testutils.RunUncoverAndGetResults(debug, "-hunter", "'Grafana'")
+	if err != nil {
+		return err
+	}
+	return expectResultsGreaterThanCount(results, 0)
+}
 
 type quakeTestcases struct{}
 
 func (h quakeTestcases) Execute() error {
-	token := os.Getenv("QUAKE_API_KEY")
+	token := os.Getenv("QUAKE_TOKEN")
 	if token == "" {
 		return errors.New("missing quake api key")
 	}
 	quakeToken := fmt.Sprintf(`quake: [%s]`, token)
-	_ = os.WriteFile(ConfigFile, []byte(quakeToken), 0644)
+	_ = os.WriteFile(ConfigFile, []byte(quakeToken), os.ModePerm)
 	defer os.RemoveAll(ConfigFile)
 	results, err := testutils.RunUncoverAndGetResults(debug, "-quake", "'Grafana'")
 	if err != nil {
@@ -136,7 +322,7 @@ func (h netlasTestcases) Execute() error {
 		return errors.New("missing netlas api key")
 	}
 	netlasToken := fmt.Sprintf(`netlas: [%s]`, token)
-	_ = os.WriteFile(ConfigFile, []byte(netlasToken), 0644)
+	_ = os.WriteFile(ConfigFile, []byte(netlasToken), os.ModePerm)
 	defer os.RemoveAll(ConfigFile)
 	results, err := testutils.RunUncoverAndGetResults(debug, "-netlas", "'Grafana'")
 	if err != nil {
@@ -153,9 +339,27 @@ func (h criminalipTestcases) Execute() error {
 		return errors.New("missing criminalip api key")
 	}
 	criminalipToken := fmt.Sprintf(`criminalip: [%s]`, token)
-	_ = os.WriteFile(ConfigFile, []byte(criminalipToken), 0644)
+	_ = os.WriteFile(ConfigFile, []byte(criminalipToken), os.ModePerm)
 	defer os.RemoveAll(ConfigFile)
 	results, err := testutils.RunUncoverAndGetResults(debug, "-criminalip", "'Grafana'")
+	if err != nil {
+		return err
+	}
+	return expectResultsGreaterThanCount(results, 0)
+}
+
+type fullhuntTestcases struct {
+}
+
+func (h fullhuntTestcases) Execute() error {
+	token := os.Getenv("FULLHUNT_API_KEY")
+	if token == "" {
+		return errors.New("missing fullhunt api key")
+	}
+	fullhuntToken := fmt.Sprintf(`fullhunt: [%s]`, token)
+	_ = os.WriteFile(ConfigFile, []byte(fullhuntToken), os.ModePerm)
+	defer os.RemoveAll(ConfigFile)
+	results, err := testutils.RunUncoverAndGetResults(debug, "-fullhunt", "baidu.com")
 	if err != nil {
 		return err
 	}
@@ -170,7 +374,7 @@ func (h hunterhowTestcases) Execute() error {
 		return errors.New("missing hunterhow api key")
 	}
 	hunterhowApiKey := fmt.Sprintf(`hunterhow: [%s]`, token)
-	_ = os.WriteFile(ConfigFile, []byte(hunterhowApiKey), 0644)
+	_ = os.WriteFile(ConfigFile, []byte(hunterhowApiKey), os.ModePerm)
 	defer os.RemoveAll(ConfigFile)
 	results, err := testutils.RunUncoverAndGetResults(debug, "-hunterhow", "'web.body=\"ElasticJob\"'")
 	if err != nil {
